+++ alidns-20150109/alibabacloud_alidns20150109/models/_update_cloud_gtm_global_alert_request.py
@@ -80,6 +80,7 @@ class UpdateCloudGtmGlobalAlertRequestAlertConfig(DaraModel):
         email_notice: bool = None,
         notice_type: str = None,
         sms_notice: bool = None,
+        threshold: int = None,
     ):
         # Specifies whether to configure DingTalk notifications. Valid values:
         # 
@@ -105,6 +106,7 @@ class UpdateCloudGtmGlobalAlertRequestAlertConfig(DaraModel):
         # 
         # Only the China site (aliyun.com) supports text message notifications.
         self.sms_notice = sms_notice
+        self.threshold = threshold
 
     def validate(self):
         pass
@@ -126,6 +128,9 @@ class UpdateCloudGtmGlobalAlertRequestAlertConfig(DaraModel):
         if self.sms_notice is not None:
             result['SmsNotice'] = self.sms_notice
 
+        if self.threshold is not None:
+            result['Threshold'] = self.threshold
+
         return result
 
     def from_map(self, m: dict = None):
@@ -141,6 +146,9 @@ class UpdateCloudGtmGlobalAlertRequestAlertConfig(DaraModel):
 
         if m.get('SmsNotice') is not None:
             self.sms_notice = m.get('SmsNotice')
+
+        if m.get('Threshold') is not None:
+            self.threshold = m.get('Threshold')
 
         return self